Remove `q` binding for quitting Sketch
Instead of using `q` to quit, which prevents the user from typing the
character `q`, the binding of ^C is now documented to close the
application.

Since this works anyways and is handled properly, documenting this
binding seems more reasonable than finding a new separate binding. This
also allows to keep the IXON capability in place.

Fixes #4.