Optimize cursor drawing performance
Instead of writing every cursor cell manually, this instead writes the
first cursor cell and then repeats it with the `CSI Ps b` escape
sequence.

This allows the terminal emulator to take care of all the repetition,
without having to pipe a bunch of data through the PTY. As a result,
this significantly improves performance especially with big cursors.

Fixes #6.