Prefill colors in dialog
Instead of starting the color dialog with an empty color, the current
color is now prefilled.

The dialog boxes also don't use the default color for the opposite color
position anymore, but instead use the color which is currently active
for that position. This allows previewing the combination of background
and foreground without having to leave the dialog.