Use bitflags for SGR event parsing
Since the SGR events are structured as bitflags, the code becomes a bit
easier to read when the different fields are parsed with the help of
bitflags rather than having magic constants everywhere.

This also fixes a bug where mouse motions without any buttons would
result in the `Down` button state.