Add line drawing mode
This adds two separate line drawing modes to sketch. The first mode
allows drawing straight horizontal/vertical lines by holding control
while trying to draw a line. This mode makes sure that a line is always
straight, even when the user doesn't manage to draw the line quite
straight.

The other mode is a box drawing mode. When holding control and clicking
somewhere, this mode is entered and a box drawing preview is shown. The
second click then selects the opposite corner of the box and draws it
using the line drawing characters.

Fixes #1.