Fix artifacts when backspacing in save dialog
This fixes an issue where pressing backspace in the save dialog could
cause the left- or rightmost column to be rendered twice, since the
dialog is shrunk without clearing the previous dialog's area.

This would only happen when the specified path was longer than the
dialog prompt, since there would be no shrinkage otherwise.

Closes #26.